# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5e8c3cd98763035aac7a90c1
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e8c3d708763035aac7a90e2
  title: Hiimat
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  /dare:
    post:
      x-skaffolder-id: 5e8c43ee8763035aac7a9182
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-resource: Dare
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Dare
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dare"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Activate:
                  type: boolean
                Description:
                  type: string
                Expiring:
                  type: integer
                Start:
                  type: integer
                title:
                  type: string
    get:
      x-skaffolder-id: 5e8c42908763035aac7a917d
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-resource: Dare
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Dare
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dare"
  "/dare/{id}":
    delete:
      x-skaffolder-id: 5e8c42908763035aac7a917c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-resource: Dare
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Dare
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e8c43ee8763035aac7a9183
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-resource: Dare
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION get
      tags:
        - Dare
      parameters:
        - name: id
          x-skaffolder-type: Number
          in: path
          description: 
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dare"
    post:
      x-skaffolder-id: 5e8c43ee8763035aac7a9184
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-resource: Dare
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Dare
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dare"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dare"
  "/daringuser/findBydare/{key}":
    get:
      x-skaffolder-id: 5e8c43ee8763035aac7a9187
      x-skaffolder-name: findBydare
      x-skaffolder-id-resource: 5e8c41a68763035aac7a9171
      x-skaffolder-resource: DaringUser
      x-skaffolder-crudAction: findBydare
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBydare
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBydare/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBydare
      tags:
        - DaringUser
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DaringUser"
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5e8c3d718763035aac7a90f6
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5e8c3d718763035aac7a90ee
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5e8c3d718763035aac7a90f5
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                photo:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5e8c3d718763035aac7a90f7
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5e8c3d718763035aac7a90f1
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e8c3d718763035aac7a90ef
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5e8c3d718763035aac7a90f0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e8c3d718763035aac7a90ec
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Dare:
      x-skaffolder-id: 5e8c409e8763035aac7a9163
      x-skaffolder-id-db: 5e8c3d708763035aac7a90e3
      x-skaffolder-id-entity: 5e8c409e8763035aac7a915c
      x-skaffolder-url: /dare
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Activate:
          type: boolean
          x-skaffolder-id-attr: 5e8c409e8763035aac7a9160
          x-skaffolder-type: Boolean
        Description:
          type: string
          x-skaffolder-id-attr: 5e8c409e8763035aac7a9161
          x-skaffolder-type: String
        Expiring:
          type: integer
          x-skaffolder-id-attr: 5e8c409e8763035aac7a915e
          x-skaffolder-type: Date
          x-skaffolder-required: true
        Start:
          type: integer
          x-skaffolder-id-attr: 5e8c409e8763035aac7a915f
          x-skaffolder-type: Date
          x-skaffolder-required: true
        title:
          type: string
          x-skaffolder-id-attr: 5e8c409e8763035aac7a915d
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        Author:
          x-skaffolder-id: 5e8c409e8763035aac7a9162
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5e8c409e8763035aac7a915c
          x-skaffolder-ent2: 5e8c3d708763035aac7a90e4
      required:
        - _id
        - Expiring
        - Start
        - title
    DareVideos:
      x-skaffolder-id: 5e8c42188763035aac7a9177
      x-skaffolder-id-db: 5e8c3d708763035aac7a90e3
      x-skaffolder-id-entity: 5e8c42188763035aac7a9174
      x-skaffolder-url: /darevideos
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        video:
          type: object
          x-skaffolder-id-attr: 5e8c42188763035aac7a9175
          x-skaffolder-type: Custom
          x-skaffolder-required: true
      x-skaffolder-relations:
        dareRef:
          x-skaffolder-id: 5e8c42188763035aac7a9176
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5e8c42188763035aac7a9174
          x-skaffolder-ent2: 5e8c41a68763035aac7a9166
      required:
        - _id
        - video
    DaringUser:
      x-skaffolder-id: 5e8c41a68763035aac7a9171
      x-skaffolder-id-db: 5e8c3d708763035aac7a90e3
      x-skaffolder-id-entity: 5e8c41a68763035aac7a9166
      x-skaffolder-url: /daringuser
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        AcceptedOn:
          type: integer
          x-skaffolder-id-attr: 5e8c41a68763035aac7a9169
          x-skaffolder-type: Date
        AssignedOn:
          type: integer
          x-skaffolder-id-attr: 5e8c41a68763035aac7a916a
          x-skaffolder-type: Date
        CompletedOn:
          type: integer
          x-skaffolder-id-attr: 5e8c41a68763035aac7a9168
          x-skaffolder-type: Date
        Status:
          type: string
          x-skaffolder-id-attr: 5e8c41a68763035aac7a9167
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - Accepted
            - Completed
            - Pending
            - Rejected
      x-skaffolder-relations:
        dare:
          x-skaffolder-id: 5e8c41a68763035aac7a916b
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e8c41a68763035aac7a9166
          x-skaffolder-ent2: 5e8c409e8763035aac7a915c
        user:
          x-skaffolder-id: 5e8c41a68763035aac7a916c
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e8c41a68763035aac7a9166
          x-skaffolder-ent2: 5e8c3d708763035aac7a90e4
      required:
        - _id
    Friend:
      x-skaffolder-id: 5e8c3fc18763035aac7a9159
      x-skaffolder-id-db: 5e8c3d708763035aac7a90e3
      x-skaffolder-id-entity: 5e8c3fc08763035aac7a9156
      x-skaffolder-url: /friend
      properties:
        _id:
          type: string
          x-skaffolder-required: true
      x-skaffolder-relations:
        friend:
          x-skaffolder-id: 5e8c3fc08763035aac7a9157
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e8c3fc08763035aac7a9156
          x-skaffolder-ent2: 5e8c3d708763035aac7a90e4
        user:
          x-skaffolder-id: 5e8c3fc08763035aac7a9158
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e8c3fc08763035aac7a9156
          x-skaffolder-ent2: 5e8c3d708763035aac7a90e4
      required:
        - _id
    User:
      x-skaffolder-id: 5e8c3d718763035aac7a90ec
      x-skaffolder-id-db: 5e8c3d708763035aac7a90e3
      x-skaffolder-id-entity: 5e8c3d708763035aac7a90e4
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90e7
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90e8
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90e6
          x-skaffolder-type: String
          x-skaffolder-required: true
        photo:
          type: string
          x-skaffolder-id-attr: 5e8c3f148763035aac7a914e
          x-skaffolder-type: String
        roles:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90ea
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90e9
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e8c3d708763035aac7a90e5
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
  x-skaffolder-db:
    - x-skaffolder-id: 5e8c3d708763035aac7a90e3
      x-skaffolder-name: Hiimat_db
  x-skaffolder-page:
    - x-skaffolder-id: 5e8c44258763035aac7a918a
      x-skaffolder-name: DareEdit
      x-skaffolder-url: /dares/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-services:
        - 5e8c43ee8763035aac7a9183
      x-skaffolder-nesteds: 
      x-skaffolder-links:
        - 5e8c42ad8763035aac7a917f
        - 5e8c42ad8763035aac7a917f
        - 5e8c44258763035aac7a918a
    - x-skaffolder-id: 5e8c42ad8763035aac7a917f
      x-skaffolder-name: DareList
      x-skaffolder-url: /dares
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5e8c409e8763035aac7a9163
      x-skaffolder-services:
        - 5e8c42908763035aac7a917d
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
  x-skaffolder-roles: 
    - x-skaffolder-id: 5e8c3d718763035aac7a90ee
      x-skaffolder-name: ADMIN